requirements.txt
django==4.2
djangorestframework>=3.15.1
requests>=2.31.0
